Moving Functions to eventwaithandles.windows
